btlp: the A0 grid and the summon scene out of their guards
BtlFxStepA0 is byte-exact with the grid row reset inside the layer loop;
its five-entry switch table now has a .rodata segment of its own, with the
word of padding behind it left to a bare rodata segment. BtlTalkSceneSummon
comes out of its guard with its ten-entry switch table emitted by the C.

BtlTalkSceneDemand reaches 99.83%, with the chosen enemy's address carried
through the scratch local and the row index taken into the response local,
and BtlPlaceFallenStep steps its row and column through an int.

+++ src/btlp/placefallen.c
@@ -143,7 +143,7 @@ INCLUDE_ASM("btlp/nonmatchings/placefallen", BtlPlaceFallen);
 #ifdef NON_MATCHING
 int BtlPlaceFallenStep(void)
 {
-    short next;
+    int next;
     int   keys;
     int   cell;
 
@@ -153,29 +153,29 @@ int BtlPlaceFallenStep(void)
         BtlSePlay(1, 0);
     }
     if (keys & PAD_UP) {
-        next = g_btl_place_row - 1;
+        next = (unsigned short)g_btl_place_row - 1;
         g_btl_place_row = next;
-        if (next < 0) {
+        if ((short)next < 0) {
             next = GRID_H - 1;
             g_btl_place_row = next;
         }
     }
     if (keys & PAD_DOWN) {
-        next = g_btl_place_row + 1;
-        next &= -(next < GRID_H);
+        next = (unsigned short)g_btl_place_row + 1;
+        next &= -((short)next < GRID_H);
         g_btl_place_row = next;
     }
     if (keys & PAD_LEFT) {
-        next = g_btl_place_col - 1;
+        next = (unsigned short)g_btl_place_col - 1;
         g_btl_place_col = next;
-        if (next < 0) {
+        if ((short)next < 0) {
             next = GRID_W - 1;
             g_btl_place_col = next;
         }
     }
     if (keys & PAD_RIGHT) {
-        next = g_btl_place_col + 1;
-        next &= -(next < GRID_W);
+        next = (unsigned short)g_btl_place_col + 1;
+        next &= -((short)next < GRID_W);
         g_btl_place_col = next;
     }
